backend/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi>=0.101.0
h11==0.14.0
httptools==0.6.4
idna==3.10
pycryptodome==3.21.0
pydantic<=2.0
PyJWT==2.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==1.4.47
starlette==0.27.0
typing_extensions==4.12.2
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
pymysql-1.1.1
